Fix warnings from building with clang
rtuartscreader/ifdhandler.c:59:67: error:
      format specifies type 'unsigned long' but the argument has
type 'DWORD'
rtuartscreader/include/rtuartscreader/reader_detail.h:22:3: error:
      redefinition of typedef 'Reader' is a C11 feature

+++ rtuartscreader/ifdhandler.c
@@ -82,7 +82,7 @@ RESPONSECODE IFDHCreateChannel(DWORD Lun, DWORD Channel) {
 
     char deviceName[256] = { 0 };
 
-    if (snprintf(deviceName, sizeof(deviceName), "/dev/pcsc/%lu", Channel) < 0) {
+    if (snprintf(deviceName, sizeof(deviceName), "/dev/pcsc/%lu", (unsigned long int)Channel) < 0) {
         LOG_CRITICAL_RETURN_IFD(IFD_COMMUNICATION_ERROR, "Failed to snprintf");
     }
